functionApp/requirements.txt
azure-functions
azure-core
azure-ai-vision
azure-ai-vision-face
pyodbc

